Fixed logo for old browsers and updated readme
The README still contained Lanyon's readme content (I forked lanyon
as the initial theme). I've removed that with a minimal description
of the purpose of this repo.

When I first tried adding the logo, I used the svg object with a fail
over image, but I never got around to exporting a png version of the
svg image, so the fail over image didn't work in older browsers. Now
that I've checked out the site on an older phone, I see the entire
navigation bar doesn't really work on older browsers. I think the Lanyon
doc mentioned this. As much as I like the minimalistic approach, I'd
like something that degrades nicer on older platforms...